docs: move `integrations` and `deployments` under `docs`
