+++ packages/docs/public/_redirects
@@ -54,6 +54,8 @@
 /qwikcity/prefetching/request-response-cache/ /docs/advanced/speculative-module-fetching/ 308
 /qwikcity/prefetching/parallelizing-network-requests/ /docs/advanced/speculative-module-fetching/ 308
 /qwikcity/* /docs/:splat 308
+/integrations/* /docs/integrations/:splat 308
+/deployments/* /docs/deployments/:splat 308
 /docs/overview /docs/ 308
 /docs/overview/ /docs/ 308
 /docs/cheat/qwik-react/ /docs/guides/qwik-react/ 308
