--- packages/docs/src/routes/deployments/menu.md
@@ -1,14 +1,0 @@
-# Deployments
-
-## Deployments
-
-- [Overview](index.mdx)
-
-## Adapters
-
-- [Azure SWA](azure-swa/index.mdx)
-- [Cloudflare Pages](cloudflare-pages/index.mdx)
-- [Google Cloud Run](gcp-cloud-run/index.mdx)
-- [Netlify Edge](netlify-edge/index.mdx)
-- [Node](node/index.mdx)
-- [Vercel Edge](vercel-edge/index.mdx)
