+++ packages/docs/src/routes/docs/deployments/azure-swa/index.mdx
