+++ packages/docs/src/routes/docs/integrations/delete-layout.tsx
@@ -1,10 +1,9 @@
 import { component$, Slot, useStyles$ } from '@builder.io/qwik';
-import { ContentNav } from '../../components/content-nav/content-nav';
-import { Footer } from '../../components/footer/footer';
-import { Header } from '../../components/header/header';
-import { OnThisPage } from '../../components/on-this-page/on-this-page';
-import { SideBar } from '../../components/sidebar/sidebar';
-import styles from '../docs/docs.css?inline';
+import { ContentNav } from '../../../components/content-nav/content-nav';
+import { Footer } from '../../../components/footer/footer';
+import { Header } from '../../../components/header/header';
+import { SideBar } from '../../../components/sidebar/sidebar';
+import styles from './docs/docs.css?inline';
 
 export default component$(() => {
   useStyles$(styles);
@@ -25,7 +24,6 @@ export default component$(() => {
           <ContentNav />
           <Footer />
         </div>
-        <OnThisPage />
       </main>
     </div>
   );
