+++ packages/docs/src/routes/docs/integrations/tailwind/index.mdx
