--- packages/docs/src/routes/integrations/menu.md
@@ -1,12 +1,0 @@
-# Integrations
-
-## Integrations
-
-- [Overview](index.mdx)
-- [Partytown](partytown/index.mdx)
-- [Playwright](playwright/index.mdx)
-- [PostCSS](postcss/index.mdx)
-- [React](react/index.mdx)
-- [Styled Vanilla Extract](styled-vanilla-extract/index.mdx)
-- [Tailwind](tailwind/index.mdx)
-- [Vitest](vitest/index.mdx)
